* add OptimizeStackSize to BLAS construction. This reduces stack size improving perf by a couple % for sponza, bistro, san miguel and other scenes with a reasonably large depth

+++ IDKEngine/Source/Bvh/BVH.cs
@@ -494,7 +494,7 @@ public class BVH : IDisposable
 
         if (true)
         {
-            float totalSAH = 0.0f;
+            double totalSAH = 0.0f;
             for (int i = start; i < start + count; i++)
             {
                 totalSAH += BLAS.ComputeGlobalSAH(GetBlas(i), buildSettings);
